готова функция modified_kw
